requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
Babel==2.16.0
flake8==7.1.1
greenlet==3.1.1
python-dotenv==1.0.1
ruff==0.6.9
sqlalchemy==2.0.35